requirements.txt
uproot
awkward
pandas
matplotlib
pyyaml
tabulate
hist[plot]
scikit-learn
plotly
imbalanced-learn
cloudpickle
xgboost
tensorflow
vector